Fix webkit cards and sections
