+++ js/src/styles/dashboard.scss
@@ -22,6 +22,12 @@
                 height: 100%;
                 max-height: 100%;
 
+                .card-body {
+                    .code-cell {
+                        height: 100%;
+                    }
+                }
+
                 .section-column {
                     // This is a column section
                     width: 0;
